Fix hash table insertion algorithm.
We cannot stop searching a hash table when a dummy bucket is found
during insertion as the target key might be later in the probe chain.
The way we can re-use dummy buckets for insertion, instead is storing
the first encountered dummy bucket's index in a variable, then re-using
it when the end of the probe chain (marked by an emopty bucket) was reached.